parseToken is own module
